move tests to format
